feat(arch): PR 3/12 — move agent_store.py to infrastructure/stores
- Move src/lyra/core/stores/agent_store.py → infrastructure/stores/
- Update all callers to import from lyra.infrastructure.stores.agent_store
- Update .importlinter ignore_imports for migrated agent_store
- Add layer exceptions for core → infrastructure imports:
  - agent_store_protocol.py: make_agent_store() factory imports implementation
  - agent_refiner.py, agent.py: TYPE_CHECKING imports for type hints
- Mark agent_store.py as migrated in infrastructure/CLAUDE.md
- Update file length exemption path for moved file

Per ADR-048, SQLite store implementations move to infrastructure layer
while protocols remain in core.stores.
