+++ .importlinter
@@ -15,7 +15,6 @@ layers =
 ignore_imports =
     # Transitional: stores pending migration to infrastructure/stores (ADR-048)
     # These stores import SqliteStore from infrastructure while awaiting migration
-    lyra.core.stores.agent_store -> lyra.infrastructure.stores.sqlite_base
     lyra.core.stores.auth_store -> lyra.infrastructure.stores.sqlite_base
     lyra.core.stores.credential_store -> lyra.infrastructure.stores.sqlite_base
     lyra.core.stores.identity_alias_store -> lyra.infrastructure.stores.sqlite_base
@@ -24,6 +23,12 @@ ignore_imports =
     lyra.core.stores.prefs_store -> lyra.infrastructure.stores.sqlite_base
     lyra.core.stores.thread_store -> lyra.infrastructure.stores.sqlite_base
     lyra.core.stores.turn_store -> lyra.infrastructure.stores.sqlite_base
+    # AgentStore migrated to infrastructure (ADR-048 PR3)
+    # make_agent_store() factory imports implementation at runtime
+    lyra.core.stores.agent_store_protocol -> lyra.infrastructure.stores.agent_store
+    # TYPE_CHECKING imports for type hints (no runtime dependency)
+    lyra.core.agent_refiner -> lyra.infrastructure.stores.agent_store
+    lyra.core.agent -> lyra.infrastructure.stores.agent_store
 
 [importlinter:contract:core-stores-no-sqlite]
 name = core/stores protocols must not import SQLite drivers
