+++ src/lyra/bootstrap/bootstrap_stores.py
@@ -17,13 +17,13 @@ from dataclasses import dataclass
 from pathlib import Path
 from typing import AsyncIterator
 
-from lyra.core.stores.agent_store import AgentStore
 from lyra.core.stores.auth_store import AuthStore
 from lyra.core.stores.credential_store import CredentialStore, LyraKeyring
 from lyra.core.stores.identity_alias_store import IdentityAliasStore
 from lyra.core.stores.message_index import MessageIndex
 from lyra.core.stores.prefs_store import PrefsStore
 from lyra.core.stores.turn_store import TurnStore
+from lyra.infrastructure.stores.agent_store import AgentStore
 
 log = logging.getLogger(__name__)
