+++ src/lyra/bootstrap/bot_agent_map.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 import logging
 
 from lyra.config import DiscordBotConfig, TelegramBotConfig
-from lyra.core.stores.agent_store import AgentStore
+from lyra.infrastructure.stores.agent_store import AgentStore
 
 log = logging.getLogger(__name__)
