+++ src/lyra/cli_agent.py
@@ -13,7 +13,7 @@ from typing import Optional
 
 import typer
 
-from lyra.core.stores.agent_store import AgentStore
+from lyra.infrastructure.stores.agent_store import AgentStore
 
 agent_app = typer.Typer(name="agent", help="Manage agent configurations.")
 _DEFAULT_TOOLS = ["Read", "Grep", "Glob", "WebFetch", "WebSearch"]
