+++ tests/core/test_agent_store_crud.py
@@ -7,7 +7,7 @@ from pathlib import Path
 import aiosqlite
 import pytest
 
-from lyra.core.stores.agent_store import AgentRow, AgentStore
+from lyra.infrastructure.stores.agent_store import AgentRow, AgentStore
 
 from .conftest import make_agent_row, make_store
