+++ tests/test_cli_wizard_validate.py
@@ -13,7 +13,7 @@ import pytest
 from typer.testing import CliRunner
 
 from lyra.cli import agent_app as app
-from lyra.core.stores.agent_store import AgentRow, AgentStore
+from lyra.infrastructure.stores.agent_store import AgentRow, AgentStore
 
 # ---------------------------------------------------------------------------
 # Helpers
